Move OTA HTTP handlers to ota_handlers.c
Splits all OTA API handler code out of ui_http_server.c into a
dedicated ota_handlers.c / ota_handlers.h pair as requested by
maintainer. ui_http_server.c now delegates route registration via
a single ota_register_handlers(server) call.
