+++ components/ui_http_server/CMakeLists.txt
@@ -1,4 +1,5 @@
 idf_component_register(SRCS "ui_http_server.c"
+                             "ota_handlers.c"
                        INCLUDE_DIRS "include"
                        REQUIRES esp_http_server settings_manager dsp_processor_settings tas5805m_settings app_update
                        PRIV_REQUIRES custom_board
