+++ components/ui_http_server/include/ota_handlers.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "esp_http_server.h"
+
+#if CONFIG_SNAPCLIENT_WEB_OTA
+void ota_register_handlers(httpd_handle_t server);
+#endif
